+++ src/benchmark/benchmark_maxpool.c
@@ -6,21 +6,21 @@
 #include "benchmark.h"
 
 __attribute__((noinline))
-int maxpool_baseline(float *a, size_t n, size_t filter_size, size_t stride, float* result);
+int maxpool_baseline(double *a, size_t n, size_t filter_size, size_t stride, double* result);
 __attribute__((noinline))
-int maxpool_ssr(float *a, size_t n, size_t filter_size, size_t stride, float* result);
+int maxpool_ssr(double *a, size_t n, size_t filter_size, size_t stride, double* result);
 __attribute__((noinline))
-int maxpool_ssr_frep(float *a, size_t n, size_t filter_size, size_t stride, float* result);
+int maxpool_ssr_frep(double *a, size_t n, size_t filter_size, size_t stride, double* result);
 
 __attribute__((noinline))
-int maxpool2d_baseline(float *a, size_t n0, size_t n1, size_t f0, size_t f1, size_t s0, size_t s1, float* result);
+int maxpool2d_baseline(double *a, size_t n0, size_t n1, size_t f0, size_t f1, size_t s0, size_t s1, double* result);
 __attribute__((noinline))
-int maxpool2d_ssr(float *a, size_t n0, size_t n1, size_t f0, size_t f1, size_t s0, size_t s1, float* result);
+int maxpool2d_ssr(double *a, size_t n0, size_t n1, size_t f0, size_t f1, size_t s0, size_t s1, double* result);
 __attribute__((noinline))
-int maxpool2d_ssr_frep(float *a, size_t n0, size_t n1, size_t f0, size_t f1, size_t s0, size_t s1, float* result);
+int maxpool2d_ssr_frep(double *a, size_t n0, size_t n1, size_t f0, size_t f1, size_t s0, size_t s1, double* result);
 
 size_t pool_output_size(size_t n, size_t filter_size, size_t stride);
-void print_pattern(float *a, size_t n, size_t filter_size, size_t stride, float* result);
+void print_pattern(double *a, size_t n, size_t filter_size, size_t stride, double* result);
 
 int main() {
     uint32_t core_idx = snrt_global_core_idx();
@@ -32,12 +32,12 @@ int main() {
         size_t input_size = (size - 1) * stride + filter_size;
 
         // Initialize the input data
-        float* x = allocate(input_size, sizeof(float));
-        float* result_ref = allocate(size, sizeof(float));
-        float* result = allocate(size, sizeof(float));
+        double* x = allocate(input_size, sizeof(double));
+        double* result_ref = allocate(size, sizeof(double));
+        double* result = allocate(size, sizeof(double));
 
         for (size_t i = 0; i < input_size; i++) {
-            x[i] = (float)i;
+            x[i] = (double)i;
         }
 
         BENCH_VO(maxpool_baseline, x, input_size, filter_size, stride, result_ref);
