+++ src/copy/copy.h
@@ -7,18 +7,18 @@
 /*
  * Naive implementation of copy. Copies n elements starting at source to target
  */
-int copy_snitch(float* source, const size_t n, float* target);
+int copy_snitch(double* source, const size_t n, double* target);
 
-int copy_baseline(float* source, const size_t n, float* target);
-int copy_ssr(float* source, const size_t n, float* target);
-int copy_ssr_frep(float* source, const size_t n, float* target);
+int copy_baseline(double* source, const size_t n, double* target);
+int copy_ssr(double* source, const size_t n, double* target);
+int copy_ssr_frep(double* source, const size_t n, double* target);
 
-int copy_parallel(float* source, const size_t n, float* target);
-int copy_ssr_parallel(float* source, const size_t n, float* target);
-int copy_ssr_frep_parallel(float* source, const size_t n, float* target);
+int copy_parallel(double* source, const size_t n, double* target);
+int copy_ssr_parallel(double* source, const size_t n, double* target);
+int copy_ssr_frep_parallel(double* source, const size_t n, double* target);
 
-int copy_omp(float* source, const size_t n, float* target);
-int copy_ssr_omp(float* source, const size_t n, float* target);
-int copy_ssr_frep_omp(float* source, const size_t n, float* target);
+int copy_omp(double* source, const size_t n, double* target);
+int copy_ssr_omp(double* source, const size_t n, double* target);
+int copy_ssr_frep_omp(double* source, const size_t n, double* target);
 
 #endif
