+++ src/dot/dot.h
@@ -10,29 +10,29 @@
  * Naive implementation of dotproduct.
  * Writes the result back into the respective pointer.
  */
-int dot_baseline(const float* a,
-                 const float* b,
+int dot_baseline(const double* a,
+                 const double* b,
                  const size_t n,
-                 float* result);
+                 double* result);
 
-int dot_ssr(const float* a,
-            const float* b,
+int dot_ssr(const double* a,
+            const double* b,
             const size_t n,
-            float* result);
+            double* result);
 
-int dot_ssr_frep(const float* a,
-                 const float* b,
+int dot_ssr_frep(const double* a,
+                 const double* b,
                  const size_t n,
-                 float* result);
+                 double* result);
 
 int ssr_dvec_dvec_dotp(const double* const vals_a,
                        const double* const vals_b,
                        const size_t len,
                        double* const res);
 
-int ssr_dvec_dvec_dotpf(const float* const vals_a,
-                        const float* const vals_b,
+int ssr_dvec_dvec_dotpf(const double* const vals_a,
+                        const double* const vals_b,
                         const size_t len,
-                        float* const res);
+                        double* const res);
 
 #endif
