+++ src/onnx/acos.h
@@ -9,8 +9,8 @@
 /*
  * Naive implementation of acos. Calculates the acos of n elements starting at arr.
  */
-int acos_baseline(float* arr, const size_t n, float* result);
-int acos_ssr(float* arr, const size_t n, float* result);
-int acos_ssr_frep(float* arr, const size_t n, float* result);
+int acos_baseline(double* arr, const size_t n, double* result);
+int acos_ssr(double* arr, const size_t n, double* result);
+int acos_ssr_frep(double* arr, const size_t n, double* result);
 
 #endif
